refactor: extract _models.py, _trajectory.py, _env_setup.py from sdk.py
Phase 1 module extraction — sdk.py goes from 1015 to 670 lines:
- _models.py: RunResult, AgentInstallError, AgentTimeoutError
- _trajectory.py: session capture, agent scraping, Gemini parsing
- _env_setup.py: Dockerfile dep staging, skills injection, DinD patching, env creation
- _AGENT_SKILL_PATHS replaced with _get_agent_skill_paths() derived from registry
- All internal imports updated to canonical paths; public API unchanged
- 2 new re-export tests verify both old and new import paths
