+++ src/benchflow/__init__.py
@@ -22,7 +22,9 @@ from benchflow.acp.session import ACPSession  # noqa: F401
 from benchflow.agents.registry import AGENTS, get_agent, list_agents, register_agent, infer_env_key_for_model, is_vertex_model  # noqa: F401
 from benchflow.job import Job, JobConfig, JobResult, RetryConfig  # noqa: F401
 from benchflow.metrics import BenchmarkMetrics, collect_metrics  # noqa: F401
-from benchflow.sdk import SDK, RunResult, AgentInstallError, AgentTimeoutError, stage_dockerfile_deps  # noqa: F401
+from benchflow._env_setup import stage_dockerfile_deps  # noqa: F401
+from benchflow._models import AgentInstallError, AgentTimeoutError, RunResult  # noqa: F401
+from benchflow.sdk import SDK  # noqa: F401
 from benchflow.skills import discover_skills, install_skill, parse_skill, SkillInfo  # noqa: F401
 from benchflow.environments import SERVICES, detect_services_from_dockerfile, build_service_hooks, register_service  # noqa: F401
 from benchflow.trajectories.otel import OTelCollector  # noqa: F401
