+++ tests/test_dep_staging.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from benchflow.sdk import stage_dockerfile_deps, _dep_local_name
+from benchflow._env_setup import stage_dockerfile_deps, _dep_local_name
 
 
 class TestDepLocalName:
